+++ ffi/tinywasm/.gitignore
@@ -1,2 +1,3 @@
 lib/
 icu4x_data.postcard
+icu_capi.wasm
