+++ ffi/tinywasm/Cargo.toml
@@ -1,0 +1,39 @@
+# This file is part of ICU4X. For terms of use, please see the file
+# called LICENSE at the top level of the ICU4X source tree
+# (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
+
+[package]
+name = "icu_capi_tinywasm"
+description = "Tiny WASM example for ICU4X"
+publish = false
+version = "0.1.0"
+authors = ["The ICU4X Project Developers"]
+edition = "2021"
+resolver = "2"
+repository = "https://github.com/unicode-org/icu4x"
+license-file = "LICENSE"
+categories = ["internationalization"]
+# Keep this in sync with other crates unless there are exceptions
+include = [
+    "src/**/*",
+    "examples/**/*",
+    "benches/**/*",
+    "tests/**/*",
+    "include/**/*",
+    "Cargo.toml",
+    "LICENSE",
+    "README.md"
+]
+
+[lib]
+crate-type = ["cdylib"]
+path = "src/lib.rs"
+
+[dependencies.icu_capi]
+version = "1.0.0"
+path = "../diplomat"
+default-features = false
+features = [
+    "buffer_provider",
+    "logging",
+]
