+++ ffi/tinywasm/README.md
@@ -1,0 +1,7 @@
+# icu_capi_tinywasm [![crates.io](https://img.shields.io/crates/v/icu_capi_tinywasm)](https://crates.io/crates/icu_capi_tinywasm)
+
+This is a example of how to build a size-optimized WASM library file using ICU4X.
+
+## More Information
+
+For more information on development, authorship, contributing etc. please visit [`ICU4X home page`](https://github.com/unicode-org/icu4x).
