+++ ffi/tinywasm/src/lib.rs
@@ -1,0 +1,7 @@
+// This file is part of ICU4X. For terms of use, please see the file
+// called LICENSE at the top level of the ICU4X source tree
+// (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
+
+//! This is a example of how to build a size-optimized WASM library file using ICU4X.
+
+extern crate icu_capi;
