+++ tools/scripts/ffi.toml
@@ -9,6 +9,7 @@ description = "Run FFI tests"
 category = "ICU4X Development"
 dependencies = [
     "cargo-make-min-version",
+    "test-tinywasm",
     "test-c",
     "test-c-tiny",
     "test-cpp",
@@ -254,6 +255,19 @@ toolchain = "${ICU4X_NIGHTLY_TOOLCHAIN}"
 install_crate = { rustup_component_name = "rust-src" }
 command = "cargo"
 args = ["build", "--package", "icu", "--target", "thumbv7m-none-eabi"]
+
+[tasks.test-tinywasm]
+description = "Test the Tiny WASM example"
+category = "ICU4X Development"
+script_runner = "@duckscript"
+script = '''
+exit_on_error true
+cd ffi/tinywasm
+
+exec --fail-on-error ./build.sh
+exec --fail-on-error ls -l
+exec --fail-on-error node tiny.mjs
+'''
 
 [tasks.diplomat-get-rev]
 description = "Get current Diplomat revision"
